finish spirit user view.
登出貌似没做，保持原样提交了。
