+++ src/pages/Account/Info/BindAccount.jsx
@@ -1,18 +1,10 @@
 import React, { Component } from "react";
-import {Col, Row} from 'antd';
 
 class InfoUser extends Component {
-    constructor() {
-        super()
-    }
     render() {
         return (
             <div>
-                <Row>
-                    <Col span={20}>
-                        嘿嘿。
-                    </Col>
-                </Row>
+                敬请期待。
             </div>
         )
     }
